sandbox: specify absolute cgroup path in runtime spec
The OCI runtime is the one responsible to create the cgroup for the
containers. It takes the control group name from the optional runtime
spec "cgroupsPath" value. As per OCI runtime specification[1], it can be
a relative or absolute path. If relative "the runtime MAY interpret the
path relative to a runtime-determined location in the cgroups hierarchy"
while if absolute "the runtime MUST take the path to be relative to the
cgroups mount point".
Since cAdvisor will parse the POD cgroup for childs control groups to
retrieve the container IDs list to query from CRI-O, we need to ensure
that the runtime create the container child cgroups under the POD
cgroup, otherwise we may be unable to let kubelet fetch metrics via
cAdvisor.
Note that this actually happens with kata-runtime v2.x shimv2 runtime.

[1] https://github.com/opencontainers/runtime-spec/blob/master/config-linux.md#cgroups-path

Signed-off-by: Francesco Giudici <fgiudici@redhat.com>